PhabricatorRepositoryCommitPublishWorker: move relevance logic (1963406) (#64)
- move repository matching check to PhabricatorRepositoryCommitPublishWorker

Prior to this change, a commit that matches a revision was still going
through code in both PhabricatorRepositoryCommitPublishWorker and
DiffusionUpdateObjectAfterCommitWorker that was making changes to the
revision other than changin the status. With this change, the revisions
will still show a link to the commit, however none of the code in
DiffusionUpdateObjectAfterCommitWorker will be triggered.